TRANSFORMER CELER/requirements.txt
# Core Dependencies
pandas>=2.1.0
openpyxl>=3.1.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0
mypy>=1.7.0
black>=23.12.0
ruff>=0.1.8

# Type stubs
pandas-stubs>=2.1.0
types-openpyxl>=3.1.0

# Optional: Database support (uncomment if needed)
# sqlalchemy>=2.0.0
# psycopg2-binary>=2.9.0

# Optional: API support (uncomment if needed)
# fastapi>=0.104.0
# uvicorn>=0.24.0
